Add how to enable virtual environment
I've added some lines on README file regarding how to start
the virtual environment on terminal. It could seems straight
forward for the majority of the audience but not for new comers.

I've also included the `env` folder into the local
`exclude` file, positioned in `.git/info/exlude`.
I do believe that is better to exclude it locally, and not in
the usual `.gitignore` file for two reasons:

* The file is generated by the developer and someone could decide
 to do not use virtual environment at all.

* Is not so common to use the local exclude file. Someone could
learn something new thank to this README line :).